feat: add support for JSON response formats in openai llm
This update introduces the ability to specify JSON response formats for the Low Level Memory operations of the OpenAI API. Users have two options, json_object and json_schema modes, with the capability to also include a customizable JSON schema for the latter.

This change includes handling response formatting in the 'Chat' and 'ChatStream' methods within the openai/llm go file. In the llm/options go file, new types and options for response formatting have been added to the 'ChatOptions' struct - including an import of the encoding/json package, a new type ResponseFormatType, and a struct ResponseFormat.

Corresponding methods to set these new options were also introduced such as 'WithJSONObjectFormat', 'WithJSONSchemaFormat', and 'WithResponseFormat'. A helper type 'JSONSchemaMarshaler' was also implemented to aid in JSON binding.